Refactored code out of the Builder to reduce its complexity
