+++ src/Builders/Traits/Fields.php
@@ -5,7 +5,6 @@ namespace LaravelDoctrine\Fluent\Builders\Traits;
 use Doctrine\DBAL\Types\Type;
 use LaravelDoctrine\Fluent\Buildable;
 use LaravelDoctrine\Fluent\Builders\Field;
-use LogicException;
 
 trait Fields
 {
@@ -23,51 +22,6 @@ trait Fields
         $this->callbackAndQueue($field, $callback);
 
         return $field;
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function increments($name, callable $callback = null)
-    {
-        if ($this->isEmbeddedClass()) {
-            throw new LogicException;
-        }
-
-        return $this->integer($name, $callback)->primary()->unsigned()->autoIncrement();
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function smallIncrements($name, callable $callback = null)
-    {
-        if ($this->isEmbeddedClass()) {
-            throw new LogicException;
-        }
-
-        return $this->smallInteger($name, $callback)->primary()->unsigned()->autoIncrement();
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function bigIncrements($name, callable $callback = null)
-    {
-        if ($this->isEmbeddedClass()) {
-            throw new LogicException;
-        }
-
-        return $this->bigInteger($name, $callback)->primary()->unsigned()->autoIncrement();
     }
 
     /**
@@ -123,39 +77,6 @@ trait Fields
     public function bigInteger($name, callable $callback = null)
     {
         return $this->field(Type::BIGINT, $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function unsignedSmallInteger($name, callable $callback = null)
-    {
-        return $this->smallInteger($name, $callback)->unsigned();
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function unsignedInteger($name, callable $callback = null)
-    {
-        return $this->integer($name, $callback)->unsigned();
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function unsignedBigInteger($name, callable $callback = null)
-    {
-        return $this->bigInteger($name, $callback)->unsigned();
     }
 
     /**
@@ -252,152 +173,9 @@ trait Fields
      *
      * @return Field
      */
-    public function date($name, callable $callback = null)
-    {
-        return $this->field(Type::DATE, $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function dateTime($name, callable $callback = null)
-    {
-        return $this->field(Type::DATETIME, $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function dateTimeTz($name, callable $callback = null)
-    {
-        return $this->field(Type::DATETIMETZ, $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function time($name, callable $callback = null)
-    {
-        return $this->field(Type::TIME, $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function carbonDateTime($name, callable $callback = null)
-    {
-        return $this->field('carbondatetime', $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function carbonDateTimeTz($name, callable $callback = null)
-    {
-        return $this->field('carbondatetimetz', $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function carbonDate($name, callable $callback = null)
-    {
-        return $this->field('carbondate', $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function carbonTime($name, callable $callback = null)
-    {
-        return $this->field('carbontime', $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function zendDate($name, callable $callback = null)
-    {
-        return $this->field('zenddate', $name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function timestamp($name, callable $callback = null)
-    {
-        return $this->carbonDateTime($name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function timestampTz($name, callable $callback = null)
-    {
-        return $this->carbonDateTimeTz($name, $callback);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
     public function binary($name, callable $callback = null)
     {
         return $this->field(Type::BINARY, $name, $callback)->nullable();
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    public function rememberToken($name = 'rememberToken', callable $callback = null)
-    {
-        return $this->string($name, $callback)->nullable()->length(100);
-    }
-
-    /**
-     * @param string        $name
-     * @param callable|null $callback
-     *
-     * @return Field
-     */
-    protected function setArray($name, callable $callback = null)
-    {
-        return $this->field(Type::TARRAY, $name, $callback);
     }
 
     /**
@@ -410,9 +188,4 @@ trait Fields
      * @param callable|null $callback
      */
     abstract protected function callbackAndQueue(Buildable $buildable, callable $callback = null);
-
-    /**
-     * @return bool
-     */
-    abstract public function isEmbeddedClass();
 }
